NVIDIA: [Config]: Ensure the TPM is available before IMA initializes
BugLink: https://bugs.launchpad.net/bugs/2037688

Set the following configs:

  CONFIG_SPI_TEGRA210_QUAD=y
  CONFIG_TCG_TIS_SPI=y

On Grace systems, the IMA driver emits the following log:

  ima: No TPM chip found, activating TPM-bypass!

This occurs because the IMA driver initializes before we are able to detect
the TPM. This will always be the case when the drivers required to
communicate with the TPM, spi_tegra210_quad and tpm_tis_spi, are built as
modules.

Having these drivers as built-ins ensures that the TPM is available before
the IMA driver initializes.

Signed-off-by: Jamie Nguyen <jamien@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>